requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn>=0.30.0,<1.0.0
python-multipart>=0.0.9,<1.0.0
google-cloud-storage>=2.18.0,<3.0.0
google-generativeai>=0.8.0,<1.0.0
python-dotenv>=1.0.0,<2.0.0
